# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * পাইকারি69bd.com wholesale marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from "./order";
import type { Product } from "./product";

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  totalRevenue: number;
  totalProducts: number;
  lowStockCount: number;
  recentOrders: Order[];
  topProducts: Product[];
}