# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * পাইকারি69bd.com wholesale marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  identifier: string;
  name?: string;
  email?: string;
  phone?: string;
  shopName?: string;
  address?: string;
  district?: string;
  createdAt?: string;
}