# lib/api-zod/src/generated/types/cartLine.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * পাইকারি69bd.com wholesale marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CartLine {
  productId: string;
  qty: number;
}