# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * পাইকারি69bd.com wholesale marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from "./address";
import type { CartLine } from "./cartLine";
import type { PaymentMethod } from "./paymentMethod";

export interface CreateOrderBody {
  items: CartLine[];
  address: Address;
  paymentMethod: PaymentMethod;
  txnRef?: string;
  note?: string;
  userIdentifier?: string;
}