# lib/api-zod/src/generated/types/heroSlide.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * পাইকারি69bd.com wholesale marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface HeroSlide {
  titleBn: string;
  subtitleBn: string;
  ctaText: string;
  ctaHref: string;
  image: string;
  accent: string;
}