# lib/api-zod/src/generated/types/homeSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * পাইকারি69bd.com wholesale marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";
import type { HeroSlide } from "./heroSlide";
import type { Product } from "./product";
import type { TrustStats } from "./trustStats";

export interface HomeSummary {
  heroSlides: HeroSlide[];
  categories: Category[];
  flashDeals: Product[];
  bestSellers: Product[];
  newArrivals: Product[];
  trustStats: TrustStats;
  flashEndsAt?: string;
}