# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * পাইকারি69bd.com wholesale marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from "./address";
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";
import type { PaymentMethod } from "./paymentMethod";

export interface Order {
  orderNo: string;
  createdAt: string;
  status: OrderStatus;
  items: OrderItem[];
  address: Address;
  paymentMethod: PaymentMethod;
  txnRef?: string;
  note?: string;
  subtotal: number;
  shipping: number;
  savings?: number;
  total: number;
  userIdentifier?: string;
}