# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * পাইকারি69bd.com wholesale marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: string;
  titleBn: string;
  image: string;
  qty: number;
  unitPrice: number;
  lineTotal: number;
}